requirements.txt
openai>=1.52.2
python-dotenv>=1.0.1
colorama>=0.4.6
qdrant-client>=1.12.0
pyyaml>=6.0.1
